feat(layout): add ContainerBuilder::gap_overlap for signed inter-child gap
SLT's analogue of ratatui Layout::spacing(-1). gap_overlap(1) overlaps
adjacent children by one cell on the main axis so two bordered panels
collapse their duplicate border edge into a shared column (row layout)
or row (column layout).

The internal gap pipeline (LayoutNode::gap, ContainerConfig::gap,
BeginContainerArgs/BeginScrollableArgs::gap, ContainerBuilder::gap) is
now i32. layout_row/layout_column/justify_offsets and the intrinsic-size
kernels use i64 intermediates with saturating_add_signed, so an overlap
larger than a child's extent saturates at 0 (no panic, no wrap).

Public gap(u32)/row_gap/col_gap/ContainerStyle APIs are unchanged
(source-compatible); only gap_overlap can store a negative value.
Shared-cell rendering is last-write-wins (no automatic junction merge
yet). _ASSERT_LAYOUT_NODE_SIZE is unaffected (i32 == 4 bytes == u32).

Adds 5 layout unit tests covering overlap width reduction, gap_overlap(0)
== gap(0), large-overlap saturation, positive-gap regression, and column
direction.

Closes #222

+++ src/context/widgets_interactive/collections.rs
@@ -523,7 +523,7 @@ impl Context {
         self.commands
             .push(Command::BeginContainer(Box::new(BeginContainerArgs {
                 direction: Direction::Row,
-                gap: cal_gap,
+                gap: cal_gap as i32,
                 align: Align::Start,
                 align_self: None,
                 justify: Justify::Start,
